Update README.md
Introducing Shop_Item.Api! This project was developed as part of an intensive course, where I enthusiastically explored algorithms and data analysis to enhance my skills in software development. Although the project may be considered basic, it represents my dedication to mastering the fundamentals and building a strong foundation in this field. I am excited to continue my journey in software development, seeking new challenges and opportunities to contribute to cutting-edge advancements in the industry.